Factor out `GComparing` Monad
Not sure we want to expose this module yet, so just kept this change
purely implementation-only by putting it under `other-modules`
