+++ dependent-sum-template/src/Data/GADT/Compare/TH.hs
@@ -7,16 +7,15 @@
 module Data.GADT.Compare.TH
     ( DeriveGEQ(..)
     , DeriveGCompare(..)
-    , GComparing, runGComparing, geq', compare'
+    , module Data.GADT.Compare.Monad
     ) where
 
-import Control.Applicative
 import Control.Monad
 import Data.Dependent.Sum
 import Data.Dependent.Sum.TH.Internal
 import Data.Functor.Identity
 import Data.GADT.Compare
-import Data.Traversable (for)
+import Data.GADT.Compare.Monad
 import Data.Type.Equality ((:~:) (..))
 import Language.Haskell.TH
 import Language.Haskell.TH.Extras
@@ -70,31 +69,6 @@ geqClause bndrs con = do
             )
         ) []
     where conName = nameOfCon con
-
--- A monad allowing gcompare to be defined in the same style as geq
-newtype GComparing a b t = GComparing (Either (GOrdering a b) t)
-
-instance Functor (GComparing a b) where fmap f (GComparing x) = GComparing (either Left (Right . f) x)
-instance Monad (GComparing a b) where
-    return = pure
-    GComparing (Left  x) >>= f = GComparing (Left x)
-    GComparing (Right x) >>= f = f x
-instance Applicative (GComparing a b) where
-    pure = GComparing . Right
-    (<*>) = ap
-
-geq' :: GCompare t => t a -> t b -> GComparing x y (a :~: b)
-geq' x y = GComparing (case gcompare x y of
-    GLT -> Left GLT
-    GEQ -> Right Refl
-    GGT -> Left GGT)
-
-compare' x y = GComparing $ case compare x y of
-    LT -> Left GLT
-    EQ -> Right ()
-    GT -> Left GGT
-
-runGComparing (GComparing x) = either id id x
 
 class DeriveGCompare t where
     deriveGCompare :: t -> Q [Dec]
